move icons to ui folder
