+++ src/components/ui/icons/IconBookmark.tsx
@@ -4,12 +4,8 @@ import type { IconBaseProps } from "./shared";
 
 interface IconBookmarkProps extends IconBaseProps {}
 
-const IconBookmark = ({
-  size = "md",
-  className = "",
-  ...rest
-}: IconBookmarkProps) => {
-  return <BiBookmark size={chooseSize(size)} className={className} {...rest} />;
+const IconBookmark = ({ size = "md", className = "", ...props }: IconBookmarkProps) => {
+  return <BiBookmark size={chooseSize(size)} className={className} {...props} />;
 };
 
 export default IconBookmark;
