+++ src/pages/index.tsx
@@ -1,10 +1,5 @@
 import Head from "next/head";
-import {
-  IconAuth,
-  IconBookmark,
-  IconCart,
-  IconMenu,
-} from "@/components/icons/";
+import { IconAuth, IconBookmark, IconCart, IconMenu } from "@/components/ui/icons";
 
 export default function Home() {
   return (
@@ -16,45 +11,52 @@ export default function Home() {
         <link rel="icon" href="/favicon.ico" />
       </Head>
 
-      <h1 className="text-4xl font-bold ">ZALS</h1>
+      <div className="pt-10" />
       <p className="text-black-50">TEXT BLACK 50</p>
       <p className="text-white-0">TEXT WHITE 0</p>
       <p className="text-error-50">TEXT ERROR 50</p>
       <p className="text-warning-50">TEXT WARNING 50</p>
       <p className="text-success-50">TEXT SUCCESS 50</p>
 
-      <p className="pt-10">
+      <div className="pt-10">
         sm:
         <hr className="pb-3" />
-        <div className="flex gap-4">
+        <p className="flex gap-4">
           <IconMenu size="sm" state="CLOSED" />
+          <IconMenu size="sm" state="OPENED" />
           <IconAuth size="sm" state="LOGIN" />
+          <IconAuth size="sm" state="LOGOUT" />
           <IconCart size="sm" />
+          <IconCart size={20} color="red" />
           <IconBookmark size="sm" />
-        </div>
-      </p>
+        </p>
+      </div>
 
-      <p className="pt-10">
+      <div className="pt-10">
         md:
         <hr className="pb-3" />
-        <div className="flex gap-4">
+        <p className="flex gap-4">
           <IconMenu size="md" state="CLOSED" />
+          <IconMenu size="md" state="OPENED" />
           <IconAuth size="md" state="LOGIN" />
+          <IconAuth size="md" state="LOGOUT" />
           <IconCart size="md" />
           <IconBookmark size="md" />
-        </div>
-      </p>
+        </p>
+      </div>
 
-      <p className="pt-10">
+      <div className="pt-10">
         lg:
         <hr className="pb-3" />
-        <div className="flex gap-4">
+        <p className="flex gap-4">
           <IconMenu size="lg" state="CLOSED" />
+          <IconMenu size="lg" state="OPENED" />
           <IconAuth size="lg" state="LOGIN" />
+          <IconAuth size="lg" state="LOGOUT" />
           <IconCart size="lg" />
           <IconBookmark size="lg" />
-        </div>
-      </p>
+        </p>
+      </div>
     </>
   );
 }
